+++ aws/scenarios/microVMs/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	microVM "github.com/DataDog/test-infra-definitions/aws/scenarios/microVMs/microVMs"
+	microVM "github.com/DataDog/test-infra-definitions/aws/scenarios/microVMs/microvms"
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
 )
