+++ aws/scenarios/microVMs/microvms/libvirt_fs.go
@@ -1,9 +1,9 @@
-package microvm
+package microvms
 
 import (
 	"fmt"
 
-	"github.com/DataDog/test-infra-definitions/aws/scenarios/microVMs/microVMs/resources"
+	"github.com/DataDog/test-infra-definitions/aws/scenarios/microVMs/microvms/resources"
 	"github.com/DataDog/test-infra-definitions/aws/scenarios/microVMs/vmconfig"
 	"github.com/DataDog/test-infra-definitions/command"
 	"github.com/DataDog/test-infra-definitions/common/namer"
