+++ aws/scenarios/microVMs/microvms/resources/amd64.go
