+++ aws/scenarios/microVMs/microvms/resources/amd64/domain.xls
