+++ aws/scenarios/microVMs/microvms/resources/default.go
