+++ aws/scenarios/microVMs/microvms/resources/default/network.xls
