+++ aws/scenarios/microVMs/microvms/resources/default/pool.xml
